Fixed types used in commit c29b402349
When going through bin/ to make sure everything was using the new MongoDB API,
I changed the $filter argument to the find method from a hash to an array, as
the MongoDB-1.6.1 documentation claims that said argument should be ordered.
It turns out that (1) the argument doesn't need to be ordered and (2) arrays
are rejected as query arguments. So I'm changing $filter back from arrays to
hashes.

delete_one *does* accepts an array as $filter, but I'm changing those back to
hashes as well for consistency (hashes are accepted for that).